split code to make it more readable
